capture packets off BatchedNetworkPeer
Two hooks, on sendPacket and _receivePacket. That is where a packet exists as
one uncompressed, unencrypted buffer, and it needs nothing from the
connection's layout, so there is no struct offset left to get wrong.

BatchedNetworkPeer is deliberately not derived from NetworkPeer here. Taking
the address of a virtual member yields a vcall thunk rather than the function
itself, and hooking a thunk sends the patched entry back through the vtable
to itself. The loop never reaches the body and dispatches by jump, so it
leaves no frame on any stack to find it by; it reads as the client hanging.

Patterns come from the function entry with branch targets and rip-relative
displacements masked, and are extended past the first unique length. A wrong
match corrupts the client, a missing one only throws, so length is cheap.

install() is the single entry point now and reports what failed to an error
window. Every silent throw so far has cost a memory dump to explain.

Capture keeps what arrives rather than a sample: a bounded deque of packets
and their bodies, the leading varint as the packet id, and client or server
from the direction it travelled.

+++ src/spyglass/overlay/pane/packet_list.cpp
@@ -41,7 +41,7 @@ void draw_packet_list(Capture &capture, const float height)
 
         ImGui::TableNextColumn();
         char number[16];
-        std::snprintf(number, sizeof(number), "%d", packet.number);
+        std::snprintf(number, sizeof(number), "%llu", static_cast<unsigned long long>(packet.number));
         if (ImGui::Selectable(number, capture.selected() == index, ImGuiSelectableFlags_SpanAllColumns)) {
             capture.select(index);
         }
@@ -53,7 +53,7 @@ void draw_packet_list(Capture &capture, const float height)
         ImGui::TableNextColumn();
         text(packet.destination);
         ImGui::TableNextColumn();
-        ImGui::Text("%d", packet.length);
+        ImGui::Text("%u", packet.length);
 
         ImGui::TableNextColumn();
         ImGui::PushStyleColor(ImGuiCol_Text, packet.bad ? kBadPacket : ImGui::GetStyleColorVec4(ImGuiCol_Text));
